chore: finish up release script
