+++ scripts/release/tsconfig.json
